Added working segue to project, and keyboard.xib for the actual keyboard
Sparrow: make the actual keyboard in the keyboard.xib view file, which
can be located in the keyboard folder once you are in Xcode. And guys,
please try to find a better way to do a collection segue, take a look
at the main storyboard and you will know why.